CDP-2159 - use metaContent query for types in UnpublishProjects, refactor ScrollableTableWithMenu so the NOOP query is no longer needed
