+++ lib/graphql/queries/package.js
@@ -216,6 +216,16 @@ export const TEAM_PACKAGES_QUERY = gql`
   ${PACKAGE_DETAILS_FRAGMENT}
 `;
 
+export const PACKAGES_META_QUERY = gql`
+  query PackagesMeta {
+    packages {
+      id
+      publishedAt
+      status
+    }
+  }
+`;
+
 export const TEAM_PACKAGES_COUNT_QUERY = gql`
   query PackagesCountByTeam( $team: String!, $searchTerm: String ) {
     packages(
