+++ lib/graphql/queries/video.js
@@ -4,7 +4,7 @@ import {
   LANGUAGE_DETAILS_FRAGMENT,
   TAG_DETAILS_FRAGMENT,
   CATEGORY_DETAILS_FRAGMENT,
-  SUPPORT_FILE_DETAILS_FRAGMENT
+  SUPPORT_FILE_DETAILS_FRAGMENT,
 } from './common';
 
 /*----------------------------------------
@@ -444,6 +444,16 @@ export const TEAM_VIDEO_PROJECTS_COUNT_QUERY = gql`
       }
     ) {
       id
+    }
+  }
+`;
+
+export const VIDEO_PROJECTS_META_QUERY = gql`
+  query VideoProjectsMeta {
+    videoProjects {
+      id
+      publishedAt
+      status
     }
   }
 `;
